Explicit default parser for BeautifulSoup
Fixes the following warning:

UserWarning: No parser was explicitly specified, so I'm using the best available HTML parser for this system
("html.parser"). This usually isn't a problem, but if you run this code on another system, or in a different virtual
environment, it may use a different parser and behave differently.

To get rid of this warning, change this:

 BeautifulSoup([your markup])

 to this:

  BeautifulSoup([your markup], "html.parser")